requirements.txt
numpy==2.2.4
pandas==2.2.3
beautifulsoup4==4.13.4
requests==2.32.4
pyyaml==6.0.2
tqdm==4.66.6
boto3==1.35.94
pydantic==2.10.5